sistema-de-pedidos-spring-boot-2 Create pedidos page

+++ src/main/java/org/ufpr/sistemapedidos/repository/PedidoRepository.java
@@ -2,13 +2,16 @@ package org.ufpr.sistemapedidos.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import org.ufpr.sistemapedidos.model.Pedido;
 
 @Repository
-public interface PedidoRepository extends JpaRepository<Pedido, Long>{
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
 
     @Query("SELECT c FROM Pedido c WHERE c.id = ?1")
     Pedido findOne(Integer pedidoID);
+
+
 
 }
